Se creó middleware de atenticación

+++ backend/src/controllers/notesController.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import noteSchema from "../models/noteSchema";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req , res, next)=>{ // este tipado le da a req, res y next el tipado Request, Responce y NextFunction de Express
     try{
-        const notes = await noteSchema.find().exec();
+        const authenticatedUserId = req.session.userId;
+        assertIsDefined(authenticatedUserId);
+        const notes = await noteSchema.find({userId: authenticatedUserId}).exec(); //al pasar el id del ususario al find(), hace que renderize la notas que fueron creadas por ese usuario
         res.status(200).json(notes);
     }catch(e){
         next(e);
@@ -15,6 +18,8 @@ export const getNotes: RequestHandler = async (req , res, next)=>{ // este tipad
 export const getNote: RequestHandler = async (req, res, next) => {
     try {
         const noteId = req.params.id;
+        const authenticatedUserId = req.session.userId;
+        assertIsDefined(authenticatedUserId);
 
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(404, "Invalid Id");
@@ -24,6 +29,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if(!note){
             throw createHttpError(404, "Note not Found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) { //valida que el usuario logueado no acceda a la nota de otro 
+            throw createHttpError(401, "You cannot access this note");
         }
 
         res.status(200).json(note);
@@ -41,12 +50,15 @@ export const createNote: RequestHandler<unknown, unknown, createNoteBody, unknow
     try {
         const title = req.body.title;
         const text = req.body.text;
+        const authenticatedUserId = req.session.userId;
+        assertIsDefined(authenticatedUserId);
 
         if(!title){
             throw createHttpError(400, "Note must hace a title");
         }
 
         const notes = await noteSchema.create({
+            userId: authenticatedUserId, //le agrega el id de usuario
             title: title,
             text: text
         });
@@ -70,6 +82,8 @@ export const updateNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody
         const noteId = req.params.id;
         const title = req.body.title;
         const text = req.body.text;
+        const authenticatedUserId = req.session.userId;
+        assertIsDefined(authenticatedUserId);
 
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(404, "Invalid Id");
@@ -85,6 +99,10 @@ export const updateNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody
             throw createHttpError(404, "Note not Found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
+        }
+
         note.title = title;
         note.text = text;
 
@@ -98,6 +116,8 @@ export const updateNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody
 export const deleteNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
     try {
         const noteId = req.params.id;
+        const authenticatedUserId = req.session.userId;
+        assertIsDefined(authenticatedUserId);
 
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(404, "Invalid Id");
@@ -107,6 +127,10 @@ export const deleteNote: RequestHandler<UpdateNoteParam, unknown, UpdateNoteBody
 
         if(!note){
             throw createHttpError(404, "Note not Found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
 
         await note.deleteOne(); // el usó remove(), pero es de una versión vieja de mongoose
